Add filter_na function
This function filters rows that have any `NA`, or all `NA` values. It does so by generating the expressions to pass to `filter_()` which means that it will work for any backend supported by dplyr.

This is somewhat related to #1367.
There are also a few stack overlow questions asking how to do this: [here](http://stackoverflow.com/questions/26665319/removing-na-in-dplyr-pipe), [here](http://stackoverflow.com/questions/28857653/removing-na-observations-with-dplyrfilter), and [here](http://stackoverflow.com/questions/22353633/filter-for-complete-cases-in-data-frame-using-dplyr-case-wise-deletion). The answers given are to use `na.omit` or `complete.cases` which don't extend to other backends supported by dplyr.